Create parachute system
